backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
pydantic
pydantic-settings
alembic
python-dotenv
email-validator